% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandwidth.R
\name{bandwidth}
\alias{bandwidth}
\title{Bandwidth calculation}
\usage{
bandwidth(x, rule)
}
\arguments{
\item{x}{numeric. The data from which the estimate is to be computed.}

\item{rule}{character. A rule to choose the bandwidth. See \code{\link[stats]{bw.nrd}}.}
}
\value{
A numeric value.
}
\description{
\code{bandwidth} computes the bandwidth to be used in the 
\code{\link[statip]{densityfun}} function.
}
