% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Coefficient of variation}
\usage{
cv(x, na_rm = FALSE, ...)
}
\arguments{
\item{x}{numeric. A numeric vector.}

\item{na_rm}{logical. Should missing values be removed before computing the coefficient of variation?}

\item{...}{Additional arguments to be passed to \code{\link{mean}()}.}
}
\value{
A numeric value, the coefficient of variation.
}
\description{
Compute the coefficient of variation of a numeric vector \code{x}, 
defined as the ratio between the standard deviation and the mean.
}
\references{
\url{https://en.wikipedia.org/wiki/Coefficient_of_variation}.
}
