% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midrange.R
\name{midrange}
\alias{midrange}
\title{Mid-range}
\usage{
midrange(x, na_rm = FALSE)
}
\arguments{
\item{x}{numeric. A numeric vector.}

\item{na_rm}{logical. Should missing values be removed before computing the mid-range?}
}
\value{
A numeric value, the mid-range.
}
\description{
Compute the mid-range of a numeric vector \code{x}, 
defined as the mean of the minimum and the maximum.
}
\references{
\url{https://en.wikipedia.org/wiki/Mid-range}.
}
