% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.default.R
\name{predict.default}
\alias{predict.default}
\title{Default model predictions}
\usage{
\method{predict}{default}(object, newdata, ...)
}
\arguments{
\item{object}{A model object, possibly empty.}

\item{newdata}{An optional data frame in which to look for variables 
with which to predict. 
If omitted, the fitted values are used.}

\item{...}{Additional arguments.}
}
\value{
A vector of predictions.
}
\description{
Default method of the \code{\link[stats]{predict}} generic 
function, which can be used when the model object is empty.
}
\examples{
stats::predict(NULL)
stats::predict(NULL, newdata = data.frame(x = 1:2, y = 2:3))

}
\seealso{
\code{\link[stats]{predict}} from package \pkg{stats}.
}
