% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableNA.R
\name{tableNA}
\alias{tableNA}
\title{Alternative Table Creation}
\usage{
tableNA(x)
}
\arguments{
\item{x}{numeric. An atomic vector or a factor.}
}
\value{
An object of class \code{"tableNA"}, which is the result of 
\code{\link{tabulate}()} with three attributes:
\itemize{
  \item \code{type_of_x}: the result of \code{typeof(x)};
  \item \code{is_factor_x}: the result of \code{is.factor(x)};
  \item \code{levels}: the result of \code{levels(x)}.
}
The number of missing values is always reported.
}
\description{
Count the occurrences of each factor level or value in a vector.
}
\examples{
tableNA(c(1,2,2,1,3))
tableNA(c(1,2,2,1,3, NA))

}
