% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trimean.R
\name{trimean}
\alias{trimean}
\title{Tukey's trimean}
\usage{
trimean(x, na_rm = FALSE, ...)
}
\arguments{
\item{x}{numeric. A numeric vector.}

\item{na_rm}{logical. Should missing values be removed before computing the trimean?}

\item{...}{Additional arguments to be passed to \code{\link[stats]{quantile}()}.}
}
\value{
A numeric value, the trimean.
}
\description{
Compute the trimean of a numeric vector \code{x}.
}
\references{
\url{https://en.wikipedia.org/wiki/Trimean}
}
