% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.loess.R, R/predict.default.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{plot}
\alias{reexports}
\alias{predict}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{graphics}{\code{\link[graphics]{plot}}}

  \item{stats}{\code{\link[stats]{predict}}}
}}

