% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.stepfun.R, R/picor.R, R/plot.loess.R,
%   R/predict.default.R, R/predict.kmeans.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{as.stepfun}
\alias{knots}
\alias{predict}
\alias{plot}
\alias{na.pass}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{graphics}{\code{\link[graphics]{plot}}}

  \item{stats}{\code{\link[stats]{as.stepfun}}, \code{\link[stats]{knots}}, \code{\link[stats]{na.pass}}, \code{\link[stats]{predict}}, \code{\link[stats]{predict}}}
}}

